feat: display URLs & preferences implementation
Added a `createDisplayUrl` function for parsing URL strings into their components and formatting them for display purposes.

This functionality was subtly integrated into the `MagicLink` email template to show the production URL instead of the deployment URL.

"~/preferences" was modified to accommodate the new URLs object, and underwent changes to more accurately infer only the input types for the `appConfig`, along with some renaming for clarity.