fix: brand images in SVR email and hero section
Updating the packages in a previous commit broke the brand images in our `MagicLink` React Email template. Transparent, monochromatic PNGs turned white, sometimes not appearing at all. My early suspicion was a package called "sharp" recommended by Next.js for Image Optimization, but after removing it the issue persisted.

For now, the workaround is to use opaque images in React Email templates. I plan to investigate further at a later time.

Re-uploaded `brandmark.png` (along with `brandmark-opaque.png`) to use for the Magic Link email as it seems to be a better fit.

Corrected the link to the brandmark in "~/utils/send-verification-request", added an inverted variation of `logo.png` and renamed the logos to more accurately describe each image, and changed the link in "~/app/page.tsx" to point to the inverted logo.

Minor symbol renaming was also done in "~/emails/magic-link".