feat: added a name identifier to `createSenderIdentity` & updated emails
Changed the system auth email from "auth@amnesia.onambrosia.app" to "no-reply@amnesia.onambrosia.app", and added a field in `createSenderIdentity` to identify the individual sender or role of the email in addition to the organization. Additionally, I renamed the parameters slightly, added name inference, and an interface for the options object.