feat: personal SMS*
Added an API route for receiving text messages to my personal Twilio number. This number will be ported to a service provider in the future, but for now it will just act as a vanity number that relays messages to and from a burner of mine. This doesn't save messages, and replying to contacts requires prefixing the message content with the contact's number and " | ".